Feature: add dummy protocol.

+++ lib/statistics.hpp
@@ -8,7 +8,7 @@ namespace spectrum {
 class Statistics {
     public:
     std::atomic<size_t> count_commit{0};
-    std::atomic<double> count_execution{0};
+    std::atomic<size_t> count_execution{0};
     std::atomic<size_t> count_latency_25ms{0};
     std::atomic<size_t> count_latency_50ms{0};
     std::atomic<size_t> count_latency_100ms{0};
